Adds mixed HTTP/HTTPS mode.
In order for walkhub to work with both HTTP and HTTPS websites,
a mixed mode is necessary (because of browser restrictions).

This commit makes walkhub to work with mixed HTTP/HTTPS mode.
If HTTP and HTTPS is enabled, the endpoints which are needed for
playing a walkthrough are available in HTTP. The other endpoints
are redirected to HTTPS. Because of this, HSTS is removed.